refactor(image-cropper-dialog): move into components folder
