+++ src/app/components/image-cropper-dialog/image-cropper-dialog.component.ts
@@ -1,15 +1,53 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, Inject, OnInit } from '@angular/core';
+import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
+import { ImageCroppedEvent } from 'ngx-image-cropper';
+
+export interface DialogData {
+  croppedImages: string[];
+}
 
 @Component({
   selector: 'app-image-cropper-dialog',
   templateUrl: './image-cropper-dialog.component.html',
-  styleUrls: ['./image-cropper-dialog.component.scss']
+  styleUrls: ['./image-cropper-dialog.component.scss'],
 })
 export class ImageCropperDialogComponent implements OnInit {
+  imageChangedEvent!: Event;
+  croppedImages: string[] = [];
+  currentIndex = 0;
 
-  constructor() { }
+  constructor(
+    public dialogRef: MatDialogRef<ImageCropperDialogComponent>,
+    @Inject(MAT_DIALOG_DATA) public data: DialogData
+  ) {}
 
-  ngOnInit(): void {
+  ngOnInit(): void {}
+
+  fileChangeEvent(event: Event): void {
+    this.imageChangedEvent = event;
   }
 
+  imageCropped(event: ImageCroppedEvent): void {
+    this.croppedImages[this.currentIndex] = event.base64 || '';
+  }
+
+  imageLoaded(): void {
+    // show cropper
+  }
+
+  cropperReady(): void {
+    // cropper ready
+  }
+
+  loadImageFailed(): void {
+    // show message
+  }
+
+  changeCurrentIndex(index: number): void {
+    this.currentIndex = index;
+  }
+
+  onClose(): void {
+    this.dialogRef.close();
+  }
 }
